# static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Watchpost",
    "description": "Agent-native monitoring service. Create monitors, track uptime, receive alerts — all via REST API. No signup required.",
    "version": "0.1.0",
    "license": { "name": "MIT" }
  },
  "servers": [{ "url": "/api/v1" }],
  "paths": {
...
    "/monitors/{id}/notifications": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "post": {
        "summary": "Add notification channel",
        "operationId": "createNotification",
        "tags": ["notifications"],
        "security": [{ "manageKey": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateNotification" } } }
        },
        "responses": {
          "200": { "description": "Notification channel created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NotificationChannel" } } } },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "get": {
        "summary": "List notification channels",
        "operationId": "listNotifications",
        "tags": ["notifications"],
        "security": [{ "manageKey": [] }],
        "responses": {
          "200": { "description": "Notification channels", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/NotificationChannel" } } } } },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/notifications/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "delete": {
        "summary": "Remove notification channel",
        "operationId": "deleteNotification",
        "tags": ["notifications"],
        "security": [{ "manageKey": [] }],
        "responses": {
          "200": { "description": "Notification deleted" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "summary": "Enable or disable a notification channel",
        "operationId": "updateNotification",
        "tags": ["notifications"],
        "security": [{ "manageKey": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "is_enabled": { "type": "boolean" },
                  "name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Notification channel updated" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/uptime-history": {
      "get": {
        "summary": "Aggregate uptime history by day",
        "operationId": "uptimeHistory",
        "tags": ["system"],
        "description": "Returns daily uptime percentages, check counts, and average response times aggregated across all monitors.",
        "parameters": [
          { "name": "days", "in": "query", "schema": { "type": "integer", "default": 30, "minimum": 1, "maximum": 90 }, "description": "Number of days of history" }
        ],
        "responses": {
          "200": {
            "description": "Daily uptime data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "date": { "type": "string", "format": "date" },
                      "uptime_pct": { "type": "number" },
                      "total_checks": { "type": "integer" },
                      "up_checks": { "type": "integer" },
                      "down_checks": { "type": "integer" },
                      "avg_response_ms": { "type": "number", "nullable": true }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/monitors/{id}/uptime-history": {
      "get": {
        "summary": "Per-monitor uptime history by day",
        "operationId": "monitorUptimeHistory",
        "tags": ["monitors"],
        "description": "Returns daily uptime percentages, check counts, and average response times for a specific monitor.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "days", "in": "query", "schema": { "type": "integer", "default": 30, "minimum": 1, "maximum": 90 }, "description": "Number of days of history" }
        ],
        "responses": {
          "200": {
            "description": "Daily uptime data for the monitor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "date": { "type": "string", "format": "date" },
                      "uptime_pct": { "type": "number" },
                      "total_checks": { "type": "integer" },
                      "up_checks": { "type": "integer" },
                      "down_checks": { "type": "integer" },
                      "avg_response_ms": { "type": "number", "nullable": true }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Monitor not found" }
        }
      }
    },
...
    "/settings": {
      "get": {
        "summary": "Get status page branding",
        "operationId": "getSettings",
        "tags": ["settings"],
        "responses": {
          "200": {
            "description": "Current branding settings",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Settings" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update status page branding",
        "operationId": "updateSettings",
        "tags": ["settings"],
        "security": [{ "adminKey": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateSettings" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated branding settings",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Settings" }
              }
            }
          },
          "403": { "description": "Invalid admin key" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "manageKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Monitor manage_key. Also accepted as X-API-Key header or ?key= query param."
      },
      "adminKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Service admin key (auto-generated on first run, printed to stdout). Used for service-level settings."
      }
    },
    "schemas": {
...
      "NotificationChannel": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "monitor_id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "channel_type": { "type": "string", "enum": ["webhook", "email"] },
          "config": { "type": "object" },
          "is_enabled": { "type": "boolean" },
          "created_at": { "type": "string" }
        }
      },
      "CreateNotification": {
        "type": "object",
        "required": ["name", "channel_type", "config"],
        "properties": {
          "name": { "type": "string" },
          "channel_type": { "type": "string", "enum": ["webhook", "email"] },
          "config": { "type": "object", "description": "For webhook: {\"url\": \"...\"}, for email: {\"address\": \"...\"}" }
        }
      },
...
      "Settings": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "nullable": true, "description": "Custom status page title" },
          "description": { "type": "string", "nullable": true, "description": "Custom status page description" },
          "logo_url": { "type": "string", "nullable": true, "description": "Custom logo URL for status page" }
        }
      },
      "UpdateSettings": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "description": "Set title (empty string clears)" },
          "description": { "type": "string", "description": "Set description (empty string clears)" },
          "logo_url": { "type": "string", "description": "Set logo URL (empty string clears)" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" }
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "Invalid manage key",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    }
  }
}